@@ -5,16 +5,43 @@
   "description": "A simple NextJS app with Plate",
   "scripts": {
     "build": "next build",
-    "dev": "next dev",
+    "dev": "concurrently \"contentlayer dev\" \"next dev\"",
     "start": "next start",
     "lint": "npx eslint . --ext .js,.jsx,.ts,.tsx",
     "lint:fix": "yarn lint --fix",
-    "typecheck": "yarn p:typecheck"
+    "typecheck": "contentlayer build && tsc --noEmit"
   },
   "dependencies": {
+    "@faker-js/faker": "^7.6.0",
     "@radix-ui/colors": "0.1.8",
-    "@radix-ui/react-dropdown-menu": "2.0.2",
-    "@radix-ui/react-select": "1.2.0",
+    "@radix-ui/react-accessible-icon": "^1.0.1",
+    "@radix-ui/react-accordion": "^1.1.0",
+    "@radix-ui/react-alert-dialog": "^1.0.2",
+    "@radix-ui/react-aspect-ratio": "^1.0.1",
+    "@radix-ui/react-avatar": "^1.0.1",
+    "@radix-ui/react-checkbox": "^1.0.1",
+    "@radix-ui/react-collapsible": "^1.0.1",
+    "@radix-ui/react-context-menu": "^2.1.1",
+    "@radix-ui/react-dialog": "^1.0.2",
+    "@radix-ui/react-dropdown-menu": "^2.0.1",
+    "@radix-ui/react-hover-card": "^1.0.3",
+    "@radix-ui/react-label": "^2.0.0",
+    "@radix-ui/react-menubar": "^1.0.0",
+    "@radix-ui/react-navigation-menu": "^1.1.1",
+    "@radix-ui/react-popover": "^1.0.2",
+    "@radix-ui/react-progress": "^1.0.1",
+    "@radix-ui/react-radio-group": "^1.1.0",
+    "@radix-ui/react-scroll-area": "^1.0.2",
+    "@radix-ui/react-select": "^1.2.0",
+    "@radix-ui/react-separator": "^1.0.1",
+    "@radix-ui/react-slider": "^1.1.0",
+    "@radix-ui/react-slot": "^1.0.1",
+    "@radix-ui/react-switch": "^1.0.1",
+    "@radix-ui/react-tabs": "^1.0.2",
+    "@radix-ui/react-toast": "^1.1.2",
+    "@radix-ui/react-toggle": "^1.0.1",
+    "@radix-ui/react-toggle-group": "^1.0.1",
+    "@radix-ui/react-tooltip": "^1.0.3",
     "@tippyjs/react": "4.2.5",
     "@udecode/plate": "*",
     "@udecode/plate-cloud": "workspace:^",
@@ -24,22 +51,50 @@
     "@udecode/plate-juice": "workspace:^",
     "@udecode/plate-selection": "workspace:^",
     "@udecode/plate-test-utils": "workspace:^",
+    "@vercel/analytics": "^1.0.1",
+    "@vercel/og": "^0.0.21",
     "change-case": "^4.1.2",
+    "class-variance-authority": "^0.4.0",
+    "clsx": "^1.2.1",
+    "cmdk": "^0.2.0",
+    "contentlayer": "0.3.0",
+    "date-fns": "^2.29.3",
     "lucide-react": "^0.220.0",
-    "next": "^12.1.6",
+    "next": "^13.4.3",
+    "next-contentlayer": "0.3.0",
+    "next-themes": "^0.2.1",
     "prism-react-renderer": "^1.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-pro-sidebar": "^0.7.1",
+    "react-wrap-balancer": "^0.4.0",
     "sass": "^1.54.0",
     "slate": "0.94.0",
     "slate-history": "0.93.0",
     "slate-hyperscript": "0.77.0",
     "slate-react": "0.94.0",
     "slate-test-utils": "1.3.2",
-    "tailwindcss": "^3.3.2"
+    "tailwind-merge": "^1.8.0",
+    "unist-builder": "3.0.0",
+    "unist-util-visit": "^4.1.1"
+  },
+  "resolutions": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
+    "@types/react-color": "^3.0.6",
     "autoprefixer": "^10.4.14",
+    "mdast-util-toc": "^6.1.0",
     "postcss": "^8.4.23",
+    "rehype": "^12.0.1",
+    "rehype-autolink-headings": "^6.1.1",
+    "rehype-pretty-code": "^0.6.0",
+    "rehype-slug": "^5.1.0",
+    "remark": "^14.0.2",
+    "remark-code-import": "^1.1.1",
+    "remark-gfm": "^3.0.1",
+    "shiki": "^0.12.1",
     "tailwindcss": "^3.3.2",
     "typescript": "4.6.4"
   },
