@@ -10,7 +10,7 @@ import {
   MARK_UNDERLINE,
 } from '@udecode/plate';
 
-import { Icons } from '@/plate/common/icons';
+import { Icons } from '@/components/icons';
 import { MarkToolbarButton } from '@/plate/toolbar/MarkToolbarButton';
 import { useMyPlateEditorRef } from '@/plate/typescript/plateTypes';
 
