@@ -1,7 +1,7 @@
 import React from 'react';
 import { ExcalidrawToolbarButton } from './ExcalidrawToolbarButton';
 
-import { Icons } from '@/plate/common/icons';
+import { Icons } from '@/components/icons';
 
 export function ExcalidrawElementToolbarButton() {
   return <ExcalidrawToolbarButton icon={<Icons.excalidraw />} />;
